Done all widom feature, add function to profile tab
