+++ src/trpc/routers/_app.ts
@@ -8,6 +8,7 @@ import { compatibilityRouter } from "./_compatibility";
 import { giftRouter } from "./_gift";
 import { horoscopeRouter } from "./_horoscope";
 import { profileRouter } from "./_profile";
+import { stardustRouter } from "./_stardust";
 import { wisdomRouter } from "./_wisdom";
 
 export const appRouter = createTRPCRouter({
@@ -29,6 +30,7 @@ export const appRouter = createTRPCRouter({
   getUserData: profileRouter,
   gifts: giftRouter,
   wisdom: wisdomRouter,
+  stardust: stardustRouter,
 });
 // export type definition of API
 export type AppRouter = typeof appRouter;
