+++ src/trpc/routers/_wisdom.ts
@@ -1,16 +1,13 @@
-import { and, count, desc, eq } from "drizzle-orm";
+import { countDistinct, desc, eq } from "drizzle-orm";
 import { z } from "zod";
 
 import { user } from "@/database/auth-schema";
 import { db } from "@/database/drizzle";
-import { startDustEnum } from "@/database/enums";
 import {
   decan,
   sign,
-  stardust,
   userDetail,
   wisdom,
-  wisdomStardust,
 } from "@/database/schema";
 
 import { createTRPCRouter, protectedProcedure } from "../init";
@@ -45,49 +42,6 @@ export const wisdomRouter = createTRPCRouter({
     }));
   }),
 
-  getStardustCountsByWisdomId: protectedProcedure
-    .input(z.object({ wisdomId: z.string() }))
-    .query(async ({ input }) => {
-      const counts = await db
-        .select({
-          type: stardust.type,
-          count: count(),
-        })
-        .from(wisdomStardust)
-        .innerJoin(stardust, eq(wisdomStardust.startDustId, stardust.id))
-        .where(eq(wisdomStardust.wisdomId, input.wisdomId))
-        .groupBy(stardust.type);
-
-      const result: Record<string, number> = {};
-      for (const c of counts) {
-        result[c.type] = Number(c.count);
-      }
-
-      return result;
-    }),
-
-  // Get user's stardust reaction for a wisdom
-  getUserStardustReaction: protectedProcedure
-    .input(z.object({ wisdomId: z.string() }))
-    .query(async ({ input, ctx }) => {
-      const existing = await db
-        .select({
-          type: stardust.type,
-        })
-        .from(wisdomStardust)
-        .innerJoin(stardust, eq(wisdomStardust.startDustId, stardust.id))
-        .where(
-          and(
-            eq(wisdomStardust.wisdomId, input.wisdomId),
-            eq(wisdomStardust.senderId, ctx.id),
-          ),
-        );
-      if (existing.length > 0) {
-        return { stardustType: existing[0].type };
-      }
-      return { stardustType: null };
-    }),
-
   // Get user's quotes
   getUserQuotes: protectedProcedure.query(async ({ ctx }) => {
     const quotes = await db
@@ -114,6 +68,17 @@ export const wisdomRouter = createTRPCRouter({
     }));
   }),
 
+  getUserQuotesCount: protectedProcedure.query(async ({ ctx }) => {
+    const quotesCount = await db
+      .select({
+        count: countDistinct(wisdom.id),
+      })
+      .from(wisdom)
+      .where(eq(wisdom.userId, ctx.id));
+
+    return quotesCount[0].count;
+  }),
+
   // Create wisdom
   createWisdom: protectedProcedure
     .input(
@@ -130,69 +95,5 @@ export const wisdomRouter = createTRPCRouter({
         return { success: true };
       }
       return { success: false, message: "Failed to create wisdom" };
-    }),
-
-  // Mutation: react (add/remove) stardust to wisdom
-  reactStardust: protectedProcedure
-    .input(
-      z.object({
-        wisdomId: z.string(),
-        stardustType: z.enum(startDustEnum.enumValues),
-      }),
-    )
-    .mutation(async ({ input, ctx }) => {
-      // Find stardust id by type
-      const [sd] = await db
-        .select({ id: stardust.id })
-        .from(stardust)
-        .where(eq(stardust.type, input.stardustType));
-      if (!sd) {
-        throw new Error("Invalid stardust type");
-      }
-      // Check if user already reacted to this wisdom (any type)
-      const existing = await db
-        .select({ startDustId: wisdomStardust.startDustId })
-        .from(wisdomStardust)
-        .where(
-          and(
-            eq(wisdomStardust.wisdomId, input.wisdomId),
-            eq(wisdomStardust.senderId, ctx.id),
-          ),
-        );
-      if (existing.length > 0) {
-        if (existing[0].startDustId === sd.id) {
-          // Same type: remove (toggle off)
-          await db.delete(wisdomStardust).where(
-            and(
-              eq(wisdomStardust.wisdomId, input.wisdomId),
-              eq(wisdomStardust.startDustId, sd.id),
-              eq(wisdomStardust.senderId, ctx.id),
-            ),
-          );
-          return { success: true, action: "removed", stardustType: null };
-        }
-        else {
-          await db.delete(wisdomStardust).where(
-            and(
-              eq(wisdomStardust.wisdomId, input.wisdomId),
-              eq(wisdomStardust.senderId, ctx.id),
-            ),
-          );
-          await db.insert(wisdomStardust).values({
-            wisdomId: input.wisdomId,
-            startDustId: sd.id,
-            senderId: ctx.id,
-          });
-          return { success: true, action: "changed", stardustType: input.stardustType };
-        }
-      }
-      else {
-        await db.insert(wisdomStardust).values({
-          wisdomId: input.wisdomId,
-          startDustId: sd.id,
-          senderId: ctx.id,
-        });
-        return { success: true, action: "added", stardustType: input.stardustType };
-      }
     }),
 });
